Move upload.rs into top level scope
