+++ src/minitests/mod.rs
@@ -420,7 +420,7 @@ fn reinterpret_register() {
 
 #[test]
 fn small_upgrade() {
-    use crate::evm::CodeUploader;
+    use crate::upload::CodeUploader;
     let mut machine = load_from_file(Path::new("upgradetests/upgrade1_old.mexe"));
     let uploader = CodeUploader::_new_from_file(Path::new("upgradetests/upgrade1_new.mexe"));
     let code_bytes = uploader._to_flat_vec();
@@ -442,7 +442,7 @@ fn small_upgrade() {
 
 #[test]
 fn small_upgrade_auto_remap() {
-    use crate::evm::CodeUploader;
+    use crate::upload::CodeUploader;
 
     let mut machine = load_from_file(Path::new("upgradetests/upgrade2_old.mexe"));
     let uploader = CodeUploader::_new_from_file(Path::new("upgradetests/upgrade2_new.mexe"));
